fix: fix windows installer service and missing flags
