+++ cmd/immudb/command/init.go
@@ -1,4 +1,3 @@
-// +build windows
 /*
 Copyright 2021 CodeNotary, Inc. All rights reserved.
 
@@ -23,81 +22,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
-
-func parseOptions() (options *server.Options, err error) {
-	dir, err := c.ResolvePath(viper.GetString("dir"), true)
-	if err != nil {
-		return options, err
-	}
-	port := viper.GetInt("port")
-	address := viper.GetString("address")
-
-	pidfile, err := c.ResolvePath(viper.GetString("pidfile"), true)
-	if err != nil {
-		return options, err
-	}
-	logfile, err := c.ResolvePath(viper.GetString("logfile"), true)
-	if err != nil {
-		return options, err
-	}
-	mtls := viper.GetBool("mtls")
-	auth := viper.GetBool("auth")
-	maxRecvMsgSize := viper.GetInt("max-recv-msg-size")
-	noHistograms := viper.GetBool("no-histograms")
-	detached := viper.GetBool("detached")
-	certificate, err := c.ResolvePath(viper.GetString("certificate"), true)
-	if err != nil {
-		return options, err
-	}
-	pkey, err := c.ResolvePath(viper.GetString("pkey"), true)
-	if err != nil {
-		return options, err
-	}
-	clientcas, err := c.ResolvePath(viper.GetString("clientcas"), true)
-	if err != nil {
-		return options, err
-	}
-
-	devMode := viper.GetBool("devmode")
-	adminPassword := viper.GetString("admin-password")
-	maintenance := viper.GetBool("maintenance")
-	signingKey := viper.GetString("signingKey")
-	synced := viper.GetBool("synced")
-	tokenExpTime := viper.GetInt("token-expiry-time")
-
-	webServer := viper.GetBool("web-server")
-	webServerPort := viper.GetInt("web-server-port")
-
-	storeOpts := server.DefaultStoreOptions().WithSynced(synced)
-
-	tlsConfig, err := setUpTLS(pkey, certificate, clientcas, mtls)
-	if err != nil {
-		return options, err
-	}
-
-	options = server.
-		DefaultOptions().
-		WithDir(dir).
-		WithPort(port).
-		WithAddress(address).
-		WithPidfile(pidfile).
-		WithLogfile(logfile).
-		WithTLS(tlsConfig).
-		WithAuth(auth).
-		WithMaxRecvMsgSize(maxRecvMsgSize).
-		WithNoHistograms(noHistograms).
-		WithDetached(detached).
-		WithDevMode(devMode).
-		WithAdminPassword(adminPassword).
-		WithMaintenance(maintenance).
-		WithSigningKey(signingKey).
-		WithStoreOptions(storeOpts).
-		WithTokenExpiryTime(tokenExpTime).
-		WithWebServer(webServer).
-		WithWebServerPort(webServerPort)
-
-	return options, nil
-}
 
 func (cl *Commandline) setupFlags(cmd *cobra.Command, options *server.Options) {
 	cmd.Flags().String("dir", options.Dir, "data folder")
